Правки

+++ subscription.php
@@ -7,7 +7,7 @@ require_once('functions.php');
 
 if (isset($_SESSION['user']) && isset($_GET['action'])) {
 
-	$sql = "SELECT id 
+	$sql = "SELECT id
       	    FROM user
             WHERE id = ?;";
     $stmt = mysqli_prepare($con, $sql);
@@ -15,9 +15,8 @@ if (isset($_SESSION['user']) && isset($_GET['action'])) {
  	mysqli_stmt_execute($stmt);
  	$res = mysqli_stmt_get_result($stmt);
  	$user_exists = mysqli_fetch_all($res,  MYSQLI_ASSOC);
- 	$user_exists = !empty($user_exists);
 
-	if (($_GET['action'] == 'subscribe') && ($user_exists)) {
+	if (isset($_GET['action']) && ($_GET['action'] == 'subscribe') && !empty($user_exists)) {
 
 	 	$sql = "INSERT INTO subscription (subscriber_id, author_id)
 	            VALUES  (?,?);";
@@ -26,7 +25,7 @@ if (isset($_SESSION['user']) && isset($_GET['action'])) {
 	 	mysqli_stmt_execute($stmt);
 
 
-	} elseif (($_GET['action'] == 'unsubscribe') && ($user_exists)) {
+	} elseif (isset($_GET['action']) && ($_GET['action'] == 'unsubscribe') && !empty($user_exists)) {
 		$sql = "DELETE FROM subscription
 	 			WHERE subscriber_id = ? AND author_id = ?;";
 	    $stmt = mysqli_prepare($con, $sql);
@@ -35,7 +34,7 @@ if (isset($_SESSION['user']) && isset($_GET['action'])) {
 	}
 
 
-	$refering_page = $_SERVER['HTTP_REFERER'];	
+	$refering_page = isset($_SERVER['HTTP_REFERER']) ? $_SERVER['HTTP_REFERER'] : "index.php";	
 	header("Location: $refering_page");
 	exit();
